// src/types/PokemonProfile.d.ts
export type PokemonProfile = {
  ability: any
  id: number
  name: string
  sprites: {
    front_default: string
    other: {
      'official-artwork': {
        front_default: string
      }
    }
  }
  types: Array<{
    slot: number
    type: {
      name: string
      url: string
    }
  }>
  abilities: Array<{
    ability: {
      name: string
      url: string
    }
    is_hidden: boolean
    slot: number
  }>
  stats: Array<{
    base_stat: number
    effort: number
    stat: {
      name: string
      url: string
    }
  }>
  weight: number
  height: number
  base_experience: number
  is_default: boolean
  order?: number
  location_area_encounters: string
  forms: Array<{
    name: string
    url: string
  }>
  cries?: {
    latest?: string
    legacy?: string
  }
  game_indices?: Array<{
    game_index: number
    version: {
      name: string
      url: string
    }
  }>
  held_items?: Array<any>
  moves?: Array<{
    move: {
      name: string
      url: string
    }
    version_group_details: Array<{
      level_learned_at: number
      move_learn_method: {
        name: string
        url: string
      }
      order: number | null
      version_group: {
        name: string
        url: string
      }
    }>
  }>
}